+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * New supported model class: `gamlss`. Thanks to Marcio Augusto Diniz.
 * `marginalmeans()` accepts a `wts` argument with values: "equal", "proportional", "cells".
+* `byfun` argument for the `predictions()` function to aggregate using different functions.
 * `by` argument accepts data frames for complex groupings.
 * `by` argument in `marginalmeans` only accepts data frames.
 * `by` argument accepts "group" to group by response level.